fix : member 관련 로직 id 자료형 모두 변경

+++ src/main/java/com/kakaoseventeen/dogwalking/dog/DogJpaRepository.java
@@ -5,5 +5,5 @@ import java.util.List;
 
 public interface DogJpaRepository extends JpaRepository<Dog, Long> {
 
-    List<Dog> findDogsByMemberId(int memberId);
+    List<Dog> findDogsByMemberId(Long memberId);
 }
